requirements.txt
openai>=1.0.0
requests>=2.31.0
googlesearch-python>=1.2.3
python-dotenv>=1.0.0

